Refactor - subsume Point into Position
