+++ lib/chukinas/geometry/position.ex
@@ -8,10 +8,27 @@ defmodule Position do
   # *** *******************************
   # *** TYPES
 
-  @type position_tuple() :: {number(), number()}
+  use TypedStruct
+
+  typedstruct enforce: true do
+    field :x, number(), default: 0
+    field :y, number(), default: 0
+  end
+
+  # *** *******************************
+  # *** NEW
+
+  def new(%{x: x, y: y}), do: new(x, y)
+  def new(x, y) do
+    %__MODULE__{x: x, y: y}
+  end
 
   # *** *******************************
   # *** API
+
+  def origin() do
+    new(0, 0)
+  end
 
   def to_tuple(%{x: x, y: y}), do: {x, y}
