+++ test/chukinas/geometry/position_test.exs
@@ -1,4 +1,4 @@
-alias Chukinas.Geometry.{PositionTest, Position, Pose, Point}
+alias Chukinas.Geometry.{PositionTest, Position, Pose}
 
 ExUnit.start()
 
@@ -6,11 +6,11 @@ defmodule PositionTest do
   use ExUnit.Case, async: true
 
   test "to tuple" do
-    assert {0, 0} = Point.new(0, 0) |> Position.to_tuple()
-    assert {0, 0} = Pose.new(0, 0, 0) |> Position.to_tuple()
+    assert {0, 0} = Position.origin() |> Position.to_tuple()
+    assert {0, 0} = Pose.origin() |> Position.to_tuple()
   end
 
   test "translate" do
-    assert {1, 1} = Point.new(0, 0) |> Position.translate({1, 1}) |> Position.to_tuple()
+    assert {1, 1} = Position.origin() |> Position.translate({1, 1}) |> Position.to_tuple()
   end
 end
